Create a parent widget which can be specialized into the current text search widget or into a widget with component contents
